refactor: bottomsheet Move BaseBottomSheet and reuse it
